+++ internal/httpserver/handlers.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/avivbaron/ads-analyzer/internal/analysis"
+	"github.com/avivbaron/ads-analyzer/internal/buildinfo"
 	"github.com/avivbaron/ads-analyzer/internal/models"
 	"github.com/avivbaron/ads-analyzer/internal/util"
 )
@@ -74,27 +75,27 @@ func (h *Handler) handleBatch(w http.ResponseWriter, r *http.Request) {
 
 	wg := sync.WaitGroup{}
 	worker := func() {
-	    defer wg.Done()
-	    for {
-	        select {
-	        case <-ctx.Done():
-	            return
-	        default:
-	        }
+		defer wg.Done()
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			default:
+			}
 
-	        idx, ok := <-jobs
-	        if !ok {
-	            return
-	        }
+			idx, ok := <-jobs
+			if !ok {
+				return
+			}
 
-	        res, err := h.analyzer.Analyze(ctx, req.Domains[idx])
+			res, err := h.analyzer.Analyze(ctx, req.Domains[idx])
 
-	        select {
-	        case out <- item{idx: idx, res: res, err: err}:
-	        case <-ctx.Done():
-	            return
-	        }
-	    }
+			select {
+			case out <- item{idx: idx, res: res, err: err}:
+			case <-ctx.Done():
+				return
+			}
+		}
 	}
 
 	wg.Add(workers)
@@ -104,13 +105,13 @@ func (h *Handler) handleBatch(w http.ResponseWriter, r *http.Request) {
 
 	go func() {
 		for i := range req.Domains {
-            select {
-            case jobs <- i:
-            case <-ctx.Done():
-                close(jobs)
-                return
-            }
-        }
+			select {
+			case jobs <- i:
+			case <-ctx.Done():
+				close(jobs)
+				return
+			}
+		}
 
 		close(jobs)
 		wg.Wait()
@@ -159,4 +160,57 @@ func writeAnalyzeErr(w http.ResponseWriter, err error) {
 		return
 	}
 	writeError(w, http.StatusBadGateway, err.Error())
+}
+
+// Readiness: verify cache roundtrip quickly
+func HandleReady(deps Deps) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		if deps.Cache == nil {
+			_ = json.NewEncoder(w).Encode(health{Status: "ready", Time: time.Now().UTC()})
+			return
+		}
+		ctx, cancel := context.WithTimeout(r.Context(), 200*time.Millisecond)
+		defer cancel()
+		type probe struct {
+			OK string `json:"ok"`
+		}
+		key := "ready:" + time.Now().Format("20060102150405.000")
+		p := probe{OK: "1"}
+		if err := deps.Cache.Set(ctx, key, p, time.Second); err != nil {
+			w.WriteHeader(http.StatusServiceUnavailable)
+			_ = json.NewEncoder(w).Encode(map[string]any{"status": "not ready", "error": err.Error()})
+			return
+		}
+		var out probe
+		hit, err := deps.Cache.Get(ctx, key, &out)
+		_ = deps.Cache.Delete(context.Background(), key)
+		if err != nil || !hit || out.OK != "1" {
+			w.WriteHeader(http.StatusServiceUnavailable)
+			_ = json.NewEncoder(w).Encode(map[string]any{"status": "not ready"})
+			return
+		}
+		_ = json.NewEncoder(w).Encode(health{Status: "ready", Time: time.Now().UTC()})
+	}
+}
+
+// Health: server health check
+func HandleHealth() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(health{Status: "ok", Time: time.Now().UTC()})
+	}
+}
+
+// Version: buildinfo(values injeceted at build time)
+func HandleVersion() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(map[string]string{
+			"version":    buildinfo.Version,
+			"commit":     buildinfo.Commit,
+			"build_time": buildinfo.BuildTime,
+			"go":         buildinfo.Go,
+		})
+	}
 }
